deduplicate PrintTree and move enc-only debug stuff to a new file
