+++ lib/jxl/modular/encoding/context_predict.h
@@ -398,46 +398,6 @@ struct PredictionResult {
   int32_t multiplier;
 };
 
-inline std::string PropertyName(size_t i) {
-  static_assert(kNumNonrefProperties == 16, "Update this function");
-  switch (i) {
-    case 0:
-      return "c";
-    case 1:
-      return "g";
-    case 2:
-      return "y";
-    case 3:
-      return "x";
-    case 4:
-      return "|N|";
-    case 5:
-      return "|W|";
-    case 6:
-      return "N";
-    case 7:
-      return "W";
-    case 8:
-      return "W-WW-NW+NWW";
-    case 9:
-      return "W+N-NW";
-    case 10:
-      return "W-NW";
-    case 11:
-      return "NW-N";
-    case 12:
-      return "N-NE";
-    case 13:
-      return "N-NN";
-    case 14:
-      return "W-WW";
-    case 15:
-      return "WGH";
-    default:
-      return "ch[" + ToString(15 - (int)i) + "]";
-  }
-}
-
 inline void InitPropsRow(
     Properties *p,
     const std::array<pixel_type, kNumStaticProperties> &static_props,
